requirements.txt
transformers
torch
numpy<2
gitpython
jsonlines
scikit-learn
Levenshtein
cdifflib
nltk